feat: move util function find_available_port to leptonai.util
it's often needed in photons that needs to run core component in a subprocess and communicate through localhost:port
