+++ leptonai/util/__init__.py
@@ -12,6 +12,7 @@ from .util import (
     asyncfy_with_semaphore,
     _is_local_url,
     _is_valid_url,
+    find_available_port,
 )
 
 from .s3cache import S3Cache
